add a test which passes, but I commented out the drs panic
