+++ app/upgrades/drs-5-mainnet/doc.go
@@ -1,0 +1,2 @@
+// Package drs5mainnet (3D) provides a 3D upgrade for the apps Nim+Mande from mainnet (rollapp-evm v2.2.1) to a 3D compatible version.
+package drs5mainnet
